moved input state updating and mode changing out of actions and into state library
